requirements.txt
langgraph
setuptools==70.0.0
numpy==1.23.5
uvicorn==0.30.1
fastapi==0.110.3
python-dotenv==1.0.0
crewai==0.51.1
langchain==0.2.15
langchain-aws==0.1.17
sqlalchemy==2.0.31
tiktoken==0.7.0
pydantic==2.8.2
durationpy==0.6
async-timeout
psycopg2-binary==2.9.9
anthropic
openpyxl
seaborn
statsmodels
tabula-py